--- extension/page-viewer/localpage.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8" />
-        <link rel="stylesheet" href="/lib/semantic-ui/semantic.min.css" />
-        <link rel="stylesheet" href="localpage.css" />
-    </head>
-    <body>
-        <div id="bar"></div>
-        <script src="/lib/browser-polyfill.js"></script>
-        <script src="localpage.js"></script>
-    </body>
-</html>
