+++ src/local-page.html
@@ -1,0 +1,7 @@
+<meta charset="utf-8" />
+<link rel="stylesheet" href="/lib/semantic-ui/semantic.min.css" />
+<link rel="stylesheet" href="local-page/local-page.css" />
+
+<div id="bar" />
+<script src="/lib/browser-polyfill.js"></script>
+<script src="/local-page/local-page.js"></script>
