+++ src/overview.html
@@ -1,7 +1,7 @@
 <meta charset="utf-8" />
 <title>📇 Memory overview</title>
-<link rel="stylesheet" href="style.css" />
+<link rel="stylesheet" href="/overview/overview.css" />
 
 <div id="app" />
 <script src="/lib/browser-polyfill.js"></script>
-<script src="overview.js"></script>
+<script src="/overview/overview.js"></script>
