+++ src/popup.html
@@ -51,4 +51,4 @@
 </div>
 
 <script src="/lib/browser-polyfill.js"></script>
-<script src="popup.js"></script>
+<script src="/popup/popup.js"></script>
